refactor(core): extract execute_delete + blocking helpers into handler/delete.rs
PR followup to 4c. Per the grandfather-clause exit criterion in
AGENTS.md (both main.rs and handler.rs ≤ 500 production lines),
handler.rs's 803 lines is the first reduction target. DELETE is the
heaviest verb — intent / commit / commit_failed two-step plus its
spawn_blocking helpers — so it carries enough mass to deserve its
own file and frees the most headroom in one move.

## What moves

`core/src/handler/delete.rs` (new, 306 lines):

- `pub(crate) async fn execute_delete` — verbatim from handler.rs.
- `BlockingSqliteError`, `AuditAppendJob`, `blocking_storage_error`,
  `world_exists_blocking`, `audit_append_blocking` — formerly
  `pub(crate)` (only because handler.rs's execute_delete needed
  them); now PRIVATE to delete.rs since they're co-located with
  their only caller.

`core/src/handler.rs` (803 → 523 lines):

- `mod delete;` + `pub(crate) use delete::execute_delete;` at top so
  callers (`handler::execute(verb=Delete, ...)` and white-box tests
  in `main.rs`) keep their import path stable. No call-site changes
  in main.rs / pipeline.rs.
- Deleted execute_delete (210 lines) + blocking helpers section
  (71 lines).
- Pruned now-dead imports: `audit`, `can_delete`, `Response`. The
  audit module was only used by audit_append_blocking; can_delete
  only by execute_delete's gate; Response only by the helpers'
  return types.

## Why this doesn't fully retire the grandfather clause

handler.rs is now 523 lines — 23 over the 500 ceiling. Extraction
PRs that move code OUT are explicitly exempt under the grandfather
clause, so this lands as-is. The next reduction step (provisional
followup): extract `execute_post` (~140 lines) into
`core/src/handler/post.rs`, which drops handler.rs to ~380 and
finally clears the bar for handler.rs.

main.rs (847 production lines) remains the larger target; "PR 4d"
will split route table + middleware + env parsing + Core::new/main()
into route.rs / middleware.rs / config.rs.

## Stats

- 2 files touched: handler.rs (-281 lines, 803 → 523),
  handler/delete.rs (+306 lines, new file).
- Net production: +25 lines (the per-file doc + import set duplicated
  across the new module boundary).
- handler.rs production after this PR: 523 (still over 500; see above).
- handler/delete.rs: 306 (well under 500).

## Test plan

- [x] `cargo test` — 124 unit tests pass (same set as PR 4c; no
      new tests; the rename keeps test imports stable via re-export).
- [x] `cargo clippy --all-targets -- -D warnings` — clean.
- [x] `cargo fmt --check` — clean.
- [x] `python sdk/tests/e2e_blackbox.py` — 233/233 checks pass.
- [ ] CI sdk-blackbox on Windows — same flake class as before;
      PR #109's CoAP timeout fix already in master.
